+++ tests/test_checksums.py
@@ -1,4 +1,3 @@
-
 '''
 Copyright (C) 2020-2026  Bryant Moscon - bmoscon@gmail.com