+++ CHANGES.md
@@ -4,6 +4,7 @@
  * Feature: issue #14 .items() iterator on book sides
  * Update: Add more unit tests to increase code coverage
  * Bugfix: reference / memory leaks
+ * Bugfix: issue #31 sortedDict iterator revamp
 
 ### 0.7.0 (2026-08-05)
  * Update: Modernize project (uv, pyproject.toml, etc)